Outfit selection now shows up in FourthDemoScene.
The issue with the colour pickers was that they were using the same material, and so they were effectively updating the same reference. Dk what happened with the shader tho, seems like creating a duplicate was a bad idea and that leaving just the original from the get-go would have sufficed.